+++ src/parsing/parse_utils.h
@@ -6,7 +6,7 @@
 /*   By: sliziard <sliziard@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/09/04 16:21:58 by sliziard          #+#    #+#             */
-/*   Updated: 2025/09/28 18:01:06 by sliziard         ###   ########.fr       */
+/*   Updated: 2025/09/30 18:25:34 by sliziard         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,6 +15,7 @@
 
 # include <stdbool.h>
 # include <stddef.h>
+# include <stdint.h>
 
 bool		ft_isspace(int c);
 bool		ft_isln_empty(const char *ln);
@@ -22,6 +23,6 @@ bool		ft_isln_empty(const char *ln);
 const char	*ft_skip_sp(const char *p);
 
 bool		ft_is_walkable(char c);
-bool		ft_isln_valid(const char *ln, size_t *err_idx);
+bool		ft_isln_valid(const char *ln, int32_t *err_idx);
 
 #endif
